+++ smarts/core/agent_interface.py
@@ -19,7 +19,7 @@
 # THE SOFTWARE.
 from dataclasses import dataclass, field, replace
 from enum import IntEnum
-from typing import Optional, Union
+from typing import Optional, Union, List, Tuple
 
 from .controllers import ActionSpaceType
 from .lidar_sensor_params import BasicLidar
@@ -130,6 +130,31 @@ class AgentType(IntEnum):
 
 
 @dataclass(frozen=True)
+class AgentsListAlive:
+    agents_list: List[str]
+    """The list of agents to check whether they are alive"""
+    minimum_agents_alive_in_list: int
+    """Triggers the agent to be done if the number of alive agents in agents_list falls below the given value"""
+
+
+@dataclass(frozen=True)
+class AgentsAliveDoneCriteria:
+    minimum_ego_agents_alive: Optional[int] = None
+    """If set, triggers the agent to be done if the total number of alive ego agents falls below the given value."""
+    minimum_total_agents_alive: Optional[int] = None
+    """If set, triggers the agent to be done if total number of alive agents falls below the given value."""
+    agent_lists_alive: Optional[List[AgentsListAlive]] = None
+    """A termination criteria based on the ids of agents. If set, triggers the agent to be done if any list of agents fails 
+    to meet its specified minimum number of alive agents.
+    Example: [
+        (['agent1', 'agent2'], 1),
+        (['agent3'], 1)
+    ]
+    This agent's done event would be triggered if both 'agent1' and 'agent2' is done *or* 'agent3' is done.
+    """
+
+
+@dataclass(frozen=True)
 class DoneCriteria:
     """Toggleable conditions on which to trigger episode end."""
 
@@ -149,6 +174,8 @@ class DoneCriteria:
     """End the episode when the agent is not moving for 60 seconds or more. To account
     for controller noise not moving means <= 1 meter of displacement within 60 seconds.
     """
+    agents_alive: Optional[AgentsAliveDoneCriteria] = None
+    """If set, triggers the ego agent to be done based on the number of active agents for multi-agent purposes."""
 
 
 @dataclass